+++ tests/testthat/helper.R
@@ -66,3 +66,7 @@ test_parse_time <- function(x, expected, format = "", ...) {
     equals = tibble::tibble(X1 = expected)
   )
 }
+
+expect_error_free <- function(object, ...) {
+  expect_error(object, regexp = NA)
+}